9_web/requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.7.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
Flask==2.3.2
fonttools==4.42.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sutil==0.0.2
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
thop==0.1.1.post2209072238
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.148       # 욜로v5 가져오는 것
urllib3==1.26.16
Werkzeug==2.3.6
yarl==1.9.2
